rod4sai_rss_function/requirements.txt
azure-functions
feedgen
pymongo
datetime
dnspython